Update rgl_base.cfg
This setting should stay default, it's only effecting a small percentage of the player base and is nausea inducing. It doesn't give a discernable advantage to the player using it and doesn't give any player fighting against an ultrawide user.

It only proves to give the user motion sickness (in my experience) and is a genuinely painful experience whilst playing in a competitive game overall.